test(epub): extract visual packaging family
